modify: move `exceptions.py` into `dev` folder
