+++ base_ui.py
@@ -7,7 +7,7 @@ import genshin
 import sentry_sdk
 
 import asset
-import exceptions
+import dev.exceptions as exceptions
 from apps.db import get_user_lang
 from apps.text_map import text_map
 from dev.models import CustomInteraction, ErrorEmbed, OriginalInfo
