+++ cogs/wish.py
@@ -22,7 +22,7 @@ from apps.genshin.checks import check_cookie
 from apps.text_map import text_map, to_ambr_top
 from base_ui import capture_exception
 from data.game.standard_characters import get_standard_characters
-from exceptions import FeatureDisabled
+from dev.exceptions import FeatureDisabled
 from ui.wish import ChooseBanner, ChooseWeapon, SetAuthKey, WishFilter
 from ui.wish.SetAuthKey import wish_import_command
 from utility.wish_paginator import WishPaginator
