+++ ui/genshin/EnkaDamageCalc.py
@@ -16,7 +16,7 @@ from apps.draw import main_funcs
 from apps.genshin import get_browser, get_character_fanarts
 from apps.text_map import text_map
 from base_ui import BaseView, EnkaView
-from exceptions import CardNotReady, NoCharacterFound
+from dev.exceptions import CardNotReady, NoCharacterFound
 from dev.models import CustomInteraction, DrawInput, ErrorEmbed
 from utility.utils import divide_chunks
 from yelan.data.GO_modes import HIT_MODE_TEXTS
