+++ ui/genshin/EnkaProfile.py
@@ -11,7 +11,7 @@ from apps.db import get_user_theme
 from apps.draw.main_funcs import draw_artifact_card
 from apps.text_map import text_map
 from base_ui import EnkaView
-from exceptions import FeatureDisabled
+from dev.exceptions import FeatureDisabled
 from ui.genshin import EnkaDamageCalc, ProfileSettings
 from ui.others.settings import CustomImage
 from utility import divide_chunks
