+++ ui/genshin/ProfileSettings.py
@@ -8,7 +8,7 @@ from apps.db.utility import create_user_settings, get_profile_ver
 from apps.genshin import edit_enka_cache, get_uid
 from apps.text_map import text_map
 from base_ui import BaseView, EnkaView
-from exceptions import UIDNotFound
+from dev.exceptions import UIDNotFound
 from dev.models import CustomInteraction, DefaultEmbed, ErrorEmbed
 from ui.genshin.EnkaDamageCalc import GoBack
